+++ Assets/Scripts/WallTileObject.cs
@@ -12,6 +12,6 @@ public class WallTileObject : TileObject
         renderOrder = 50;
     }
 
-    public override int OnCommand(string command, int prev) { return prev; }
-    public override void OnCommandFinished(int newValue) { }
+    public override float OnCommand(string command, float prev) { return prev; }
+    public override void OnCommandFinished(float newValue) { }
 }
